fix: db carrito

+++ src/models/cart.js
@@ -1,7 +1,7 @@
 const fs = require('fs');
 const path = require('path');
 const db = require('../database/models')
-const {Op} = require('sequelize')
+const {Op, Association} = require('sequelize')
 
 const products = require('./products')
 const users = require('./user');
@@ -40,11 +40,10 @@ module.exports = {
             if (checkStock instanceof(Error)) throw Error(checkStock)
             else { 
                 const [addProduct, created] = await db.Cart.findOrCreate({
-                    where: [
-                        {product_id: +body.id},
-                        {color_id: +body.color},
-                        {user_id: +id}
-                    ],
+                    where: {
+                        product_color_id: checkStock.id,
+                        user_id: +id
+                    },
                     defaults: {quantity: +body.cantidad},
                     logging: false
                 })
@@ -55,6 +54,7 @@ module.exports = {
                 } else throw new Error(`el articulo ${checkStock.product_id} y color ${checkStock.color_id} ya existen en el carrito`)
             }
         } catch (error) {
+            console.log(error)
             return error
         }
     },
@@ -64,16 +64,18 @@ module.exports = {
             const checkStock = await this.checkStock(body.id, body.color, body.cantidad)
             if (checkStock instanceof(Error)) throw Error(checkStock)
             else {
-                const updCart = await db.Cart.findOne({
-                    where: [
-                        {product_id: +body.id},
-                        {color_id: +body.color},
-                        {user_id: +id}
-                    ],
+                const updCart = await db.Cart.update({quantity: +body.cantidad},{
+                    where: {
+                        product_color_id: checkStock.id,
+                        user_id: +id
+                    },
                     logging: false})
-                if (updCart && +updCart.quantity != +body.cantidad) {
-                    updCart.quantity = +body.cantidad
-                    updCart.save()
+                if (updCart) {
+                    // console.log(updCart)
+                    // console.log(+updCart.quantity,  typeof(+updCart.quantity))
+                    // console.log(+body.cantidad, typeof(+body.cantidad) )
+                    // updCart.quantity = +body.cantidad
+                    // updCart.save()
                     return {success: true, message: 'Carrito actualizado'}
                 } else if (updCart) return {success: true, message: 'carrito sin cambios'}
                 else throw Error(`No se puedo actualizar la cantidad ${body.cantidad} para el producto: ${body.id} color: ${body.color}`)
@@ -86,47 +88,34 @@ module.exports = {
     delete: async function (data) {
         try {
             const {body, id} = data
+            const checkStock = await this.checkStock(body.id, body.color, 1)
+            if (checkStock instanceof(Error)) throw Error(checkStock)
             const deleteProduct = await db.Cart.destroy({
                 where: {
-                    product_id: body.id,
-                    user_id: id,
-                    color_id: body.color
-                },logging: false})
+                    product_color_id: +checkStock.id,
+                    user_id: +id,
+                }
+                //,logging: false
+            })
             if (deleteProduct >= 1) return {success: true, message: "producto borrado con exito"}
             else throw new Error(`No se pudo borrar el producto: ${body.id}, color: ${body.color}`)
         } catch (error) {
+            console.log(error)
             return error
         }
     },
     detail: async function (userid) {
         try {
-            const data = await db.Cart.findAll({
-                include: [
-                    {
-                        association: 'product',
-                        attributes: ['id','name','price'],
-                        include: {
-                            association: 'images',
-                            attributes: ['pathName'],
-                            through: {attributes: []}
-                        }
-                    }
-                    ,{
-                        association: 'color',
-                        attributes: ['id','name_es','hex'],
-                        include: {
-                            association: 'available',
-                            attributes: ['stock']
-                        }
-                    }
-            ],
-                where: {
-                    user_id: +userid
-                },
-                attributes: ['quantity'],
-                logging: false
-            })
-            return data
+            const [results, metadata] = await db.sequelize.query(`select JSON_OBJECT('id', p.id, 'name', p.name,'price', p.price, 'images', JSON_ARRAYAGG(JSON_OBJECT('id', i.id,'pathName', i.pathName))) as product, 
+            JSON_OBJECT('id',c.id ,'name', c.name, 'name_es', c.name_es, 'hex', c.hex) as color, cart.quantity as quantity,pc.stock as stock from cart
+            inner join product_colors pc on cart.product_color_id = pc.id
+            inner join products p on p.id = pc.product_id
+            inner join colors c on c.id = pc.color_id
+            inner join prod_images pi on pi.product_id = p.id
+            inner join images i on i.id = pi.image_id
+            where cart.user_id = ${userid}
+            group by pc.id;`,{logging: false})
+            return results
         } catch (error) {
             return error
         }
